+++ example/lib/widgets/other_utils/bottom/catalog/preview/warning_info_widget.preview.dart
@@ -47,7 +47,6 @@ class WarningInfoWidgetPreview extends ParentPreviewWidget {
                     clipBehavior: Clip.hardEdge,
                     child: Container(
                       padding: const EdgeInsets.all(15),
-                      color: Colors.white,
                       child: Row(
                         children: [
                           Expanded(
@@ -79,7 +78,6 @@ class WarningInfoWidgetPreview extends ParentPreviewWidget {
                     clipBehavior: Clip.hardEdge,
                     child: Container(
                       padding: const EdgeInsets.all(15),
-                      color: Colors.white,
                       child: Row(
                         children: [
                           Expanded(