+++ lib/src/builders/preview/grid_color.dart
@@ -1,0 +1,8 @@
+import 'dart:ui';
+
+class GridColor {
+  final Color light;
+  final Color dark;
+
+  GridColor({required this.light, required this.dark});
+}